+++ changelog.md
@@ -1,5 +1,11 @@
 
 
+## Roll protocol to r891108
+###### _2021-06-09 23:16:17_ | Diff: [cbc2ddb...28c241d](https://github.com/ChromeDevTools/devtools-protocol/compare/cbc2ddb...28c241d)
+#### `Network`: removed command
+* [`Network.setDataSizeLimitsForTest`](https://chromedevtools.github.io/devtools-protocol/tot/Network/#method-setDataSizeLimitsForTest)
+
+
 ## Roll protocol to r890975
 ###### _2021-06-09 15:17:50_ | Diff: [bfcd0a3...cbc2ddb](https://github.com/ChromeDevTools/devtools-protocol/compare/bfcd0a3...cbc2ddb)
 #### `CSS`: new command